파일분리완료 ( 20.12.09)
리스트 외의 코드 common폴더로  파일분리  완료
