+++ TeamProject/static/js/board/main.js
@@ -2,7 +2,7 @@ import * as FETCH from "./fetch.js";
 import * as REND from "./render.js";
 import * as REND_LIST from "./list/render.js"
 import * as FETCH_LIST from "./list/fetch.js"
-import * as EVENT from "./event.js"
+import * as EVENT from "./common/event.js"
 /*
   BOARD = 게시판
   POST = 게시글, 특히 전체조회, 포스트는 20개단위로 페이징 되고 , 맨아래로 내렸을때 다음페이지를 로드함
@@ -14,18 +14,18 @@ import * as EVENT from "./event.js"
   location.href로 링크 이동을하면 hash change이벤트가 발생하여 router.js의 router함수가 실행됨
 */
 
-//tag 생성기 , tag = tag명 A = 속성 ,B = 속성에 들어갈 내용 , C= textNode
-export const create_html_object = (tag, attribute, content, text) => {
-  const object = document.createElement(`${tag}`);
-  for (var i = 0; i <= attribute.length - 1; i++) {
-    object.setAttribute(`${attribute[i]}`, `${content[i]}`);
-  }
-  if (text !== undefined) {
-    const textNode = document.createTextNode(`${text}`);
-    object.appendChild(textNode);
-  }
-  return object;
-}
+// //tag 생성기 , tag = tag명 A = 속성 ,B = 속성에 들어갈 내용 , C= textNode
+// export const create_html_object = (tag, attribute, content, text) => {
+//   const object = document.createElement(`${tag}`);
+//   for (var i = 0; i <= attribute.length - 1; i++) {
+//     object.setAttribute(`${attribute[i]}`, `${content[i]}`);
+//   }
+//   if (text !== undefined) {
+//     const textNode = document.createTextNode(`${text}`);
+//     object.appendChild(textNode);
+//   }
+//   return object;
+// }
 
 //============입력창 클릭시 크게만들어주는 함수===================
 //재민 part
@@ -126,17 +126,18 @@ export function validFileType(file) {
 
 //서버에서 받아온 날짜를 가공해서 반환
 //재민 part
-export function calc_date(cur_date) {
-  const cur_date_list = cur_date.split(' ');
-  const month = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
-  let cur_month;
-  month.forEach((e, index) => {
-    if (cur_date_list[2] == e) cur_month = index + 1;
-  });
+// export function calc_date(cur_date) {
+//   const cur_date_list = cur_date.split(' ');
+//   const month = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+//   let cur_month;
+//   month.forEach((e, index) => {
+//     if (cur_date_list[2] == e) cur_month = index + 1;
+//   });
 
-  const date = `${cur_date_list[3]}년 ${cur_month}월 ${cur_date_list[1]}일 `;
-  return date;
-}
+//   const date = `${cur_date_list[3]}년 ${cur_month}월 ${cur_date_list[1]}일 `;
+//   return date;
+// }
+//이거 그냥 common으로 옮길게요!!
 
 /*=============좋아요 추가하기 ============*/
 //재민 part
@@ -331,51 +332,3 @@ export function calc_date(cur_date) {
 // }
 
 //export const INPUT_DATA_FILE = new img_file_hub();
-//===========나연 남길거============//
-
-/*=============================사이드바 =========================*/
-// 베스트 게시글 불러오기
-export async function loading_best_post() {
-  try {
-    const board_id = location.hash.split('#')[1];
-    const data = await FETCH.get_best_post_information(board_id);
-    if (data !== null) {
-      REND.best_post(data);
-    }
-  } catch (error) {
-    console.log(error);
-  }
-}
-export async function loading_board_information(hashValue) {
-  let board_information;
-
-  //현재 전체검색이 아닌경우 보드정보를 불러오고 전체검색인경우 보드정보를 직접만듬
-  if (hashValue[1] !== 'total') {
-    await FETCH_LIST.get_Board(hashValue[1]).then((result) => {
-      board_information = result;
-    })
-  } else board_information = {
-    board_name: '전체',
-    id: null //값 바꾸기 
-  };
-  return board_information;
-}
-// 보드정보 불러오는 코드 매서드 추출
-//==========검색기능 이벤트===========//
-export function search_function() {
-  EVENT.attach_event_when_search(document.querySelector('.side_search'), 'side');
-  EVENT.attach_event_when_search(document.querySelector('.search_bar'), 'total');
-};
-// 검색결과를 랜더링 해주는 함수
-export const loading_search_results_posts = async (hashValue, json) => { //render_searchResult()
-  const data = json.returnlist;
-  REND.title_and_side_setting(hashValue);
-
-  if (hashValue[1] === 'total') { //전체게시판 검색일경우
-    document.querySelectorAll('.post_board').forEach(item => item.style.cssText = 'display : block');
-    await REND_LIST.post_list(data, 'total'); //전체검색결과를 그린다는 확인 flag 'total'
-  } else {
-    REND_LIST.post_list(data); //일반적 검색결과
-  }
-}
-//전체 검색일때랑 사이드 검색일때 메서드 추출 (다른 곳 중복된 곳 있는지 확인해보기)
